feat: Added download pdf and refactor
