+++ client/package.json
@@ -16,6 +16,7 @@
     "@types/react-router-dom": "^5.3.3",
     "framer-motion": "^12.12.1",
     "html2canvas": "^1.4.1",
+    "html2canvas-pro": "^1.5.11",
     "jspdf": "^3.0.1",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
