+++ client/src/components/Homepage/CallToAction.tsx
@@ -1,10 +1,7 @@
 import React from "react";
+import { Link } from "react-router";
 
-interface CallToActionProps {
-  scrollToDemo: () => void;
-}
-
-const CallToAction: React.FC<CallToActionProps> = ({ scrollToDemo }) => {
+const CallToAction: React.FC = () => {
   return (
     <section className="w-full py-20 px-4 bg-blue-600 text-white text-center">
       <div className="container mx-auto">
@@ -15,15 +12,12 @@ const CallToAction: React.FC<CallToActionProps> = ({ scrollToDemo }) => {
           Try the live demo now and see how easy it is to create tailored
           questions for any role.
         </p>
-        {/* CTA Button - will likely use scrolling later */}
-        <button
-          onClick={scrollToDemo}
+        <Link
+          to={"/generate-questions"}
           className="bg-white text-blue-600 font-semibold py-3 px-8 rounded-full text-lg shadow-lg hover:bg-gray-200 transition duration-300 ease-in-out transform hover:scale-105"
         >
           Start Generating Now
-        </button>
-        {/* Optional: Support Email */}
-        {/* <p className="mt-8 text-gray-200">Need help? Contact us at support@example.com</p> */}
+        </Link>
       </div>
     </section>
   );
