+++ client/src/components/Homepage/Testimonials.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { FiUser, FiUserCheck, FiUsers } from "react-icons/fi";
 
 const Testimonials: React.FC = () => {
   const testimonials = [
@@ -6,19 +7,19 @@ const Testimonials: React.FC = () => {
       quote:
         "This AI interview question generator saved me hours! The questions were spot-on and helped me find the perfect candidate.",
       name: "Jane Doe, Senior Recruiter",
-      avatar: "https://via.placeholder.com/150/a0a0a0/ffffff?text=JD", // Placeholder avatar
+      avatar: <FiUser className="h-12 w-12" />, // Placeholder avatar
     },
     {
       quote:
         "A fantastic tool for technical hiring managers. It quickly generates relevant questions tailored to the role.",
       name: "John Smith, Engineering Lead",
-      avatar: "https://via.placeholder.com/150/b0b0b0/ffffff?text=JS", // Placeholder avatar
+      avatar: <FiUserCheck className="h-12 w-12" />, // Placeholder avatar
     },
     {
       quote:
         "Easy to use and highly effective. Gets the job done and provides great question ideas.",
       name: "Alice Johnson, HR Manager",
-      avatar: "https://via.placeholder.com/150/c0c0c0/ffffff?text=AJ", // Placeholder avatar
+      avatar: <FiUsers className="h-12 w-12" />, // Placeholder avatar
     },
   ];
 
@@ -32,13 +33,9 @@ const Testimonials: React.FC = () => {
           {testimonials.map((testimonial, index) => (
             <div
               key={index}
-              className="bg-gray-100/70 backdrop-blur-md rounded-lg shadow-lg p-6 flex flex-col items-center text-center border border-gray-200/50"
+              className="bg-gray-100/70 space-y-4 backdrop-blur-md rounded-lg shadow-lg p-6 flex flex-col items-center text-center border border-gray-200/50"
             >
-              <img
-                src={testimonial.avatar}
-                alt={`${testimonial.name}'s avatar`}
-                className="w-20 h-20 rounded-full mb-4 shadow-md"
-              />
+              {testimonial.avatar}
               <p className="text-gray-700 italic mb-4">"{testimonial.quote}"</p>
               <p className="text-gray-800 font-semibold">
                 - {testimonial.name}
