+++ client/src/components/shared/Footer.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Footer: React.FC = () => {
+  return (
+    <footer className="w-full py-12 px-4 bg-gray-800 text-gray-300 text-center">
+      <p className="mb-4 md:mb-0">
+        © 2023 AI Interview Question Generator. All rights reserved.
+      </p>
+    </footer>
+  );
+};
+
+export default Footer;
